16: Added functionality for routers tested with TDD
