+++ __tests__/app.test.js
@@ -520,7 +520,7 @@ describe("DELETE  /api/comments/:comment_id", () => {
 describe("DELETE  /api/comments/:comment_id", () => {
   test("400: Responds with 'Bad Request' error message when delete path is invalid", () => {
     return request(app)
-      .delete("/api/comments/notacommenid")
+      .delete("/api/comments/notacommentid")
       .expect(400)
       .then(({ body: { msg } }) => {
         expect(msg).toBe("Bad Request");
