+++ controllers/api_controllers.js
@@ -1,4 +1,3 @@
-// const { fetchApi } = require("../models/api_models.js");
 const endpoints = require("../endpoints.json");
 
 exports.getApiEndpoints = (req, res, next) => {
