+++ controllers/article_controllers.js
@@ -2,6 +2,8 @@ const {
   fetchArticleById,
   updateArticleById,
   fetchAllArticles,
+  fetchCommentsByArticleId,
+  insertCommentByArticleId,
 } = require("../models/article_models.js");
 
 // Trello 4
@@ -33,6 +35,36 @@ exports.getAllArticles = (req, res, next) => {
   fetchAllArticles(sort_by, order, topic)
     .then((articles) => {
       res.status(200).send({ articles });
+    })
+    .catch((err) => {
+      next(err);
+    });
+};
+// Trello 9
+exports.getCommentsByArticleId = (req, res, next) => {
+  // const params = req.params;
+  const { article_id } = req.params;
+
+  const promises = [
+    fetchArticleById(article_id),
+    fetchCommentsByArticleId(article_id),
+  ];
+
+  Promise.all(promises)
+    .then(([articleObject, comments]) => {
+      res.status(200).send({ comments });
+    })
+    .catch((err) => {
+      next(err);
+    });
+};
+exports.postCommentByArticleId = (req, res, next) => {
+  const { article_id } = req.params;
+  const { username, body } = req.body;
+
+  insertCommentByArticleId(article_id, username, body)
+    .then((postedComment) => {
+      res.status(201).send({ postedComment });
     })
     .catch((err) => {
       next(err);
