+++ package.json
@@ -32,6 +32,7 @@
     "supertest": "^6.2.4"
   },
   "dependencies": {
+    "cors": "^2.8.5",
     "dotenv": "^16.0.1",
     "express": "^4.18.1",
     "pg": "^8.7.3"
