+++ routes/api_router.js
@@ -1,0 +1,6 @@
+const apiRouter = require("express").Router();
+const { getApiEndpoints } = require("../controllers/api_controllers");
+
+apiRouter.get("/", getApiEndpoints);
+
+module.exports = apiRouter;
