+++ routes/topic_router.js
@@ -1,0 +1,5 @@
+const topicsRouter = require("express").Router();
+const { getAllTopics } = require("../controllers/topic_controllers");
+topicsRouter.get("/", getAllTopics);
+
+module.exports = topicsRouter;
