+++ routes/user_routers.js
@@ -1,0 +1,7 @@
+const usersRouter = require("express").Router();
+
+const { getAllUsers } = require("../controllers/user_controller.js");
+
+usersRouter.get("/", getAllUsers);
+
+module.exports = usersRouter;
